Crate lmdb

Idiomatic and safe APIs for interacting with the Lightning Memory-mapped Database (LMDB).

Structs

Database

A handle to an individual database in an environment.

DatabaseFlags

Database options.

Environment

An LMDB environment.

EnvironmentBuilder

Options for opening or creating an environment.

EnvironmentFlags

Environment options.

InactiveTransaction

An inactive read-only transaction.

Info

Environment information.

RoCursor

A read-only cursor for navigating the items within a database.

RoTransaction

An LMDB read-only transaction.

RwCursor

A read-write cursor for navigating items within a database.

RwTransaction

An LMDB read-write transaction.

Stat

Environment statistics.

WriteFlags

Write options.

Enums

Error

An LMDB error kind.

Iter

An iterator over the key/value pairs in an LMDB database.

IterDup

An iterator over the keys and duplicate values in an LMDB database.

Traits

Cursor

An LMDB cursor.

Transaction

An LMDB transaction.

Type Aliases

Result

An LMDB result.