Struct lmdb::core::CursorUnstable

pub struct Cursor<'a> {
    // some fields omitted
}

Methods

impl<'a> Cursor<'a>

fn to_first(&mut self) -> MdbResult<()>

Moves cursor to first entry

fn to_last(&mut self) -> MdbResult<()>

Moves cursor to last entry

fn to_key<'k, K: ToMdbValue<'k>>(&mut self, key: &'k K) -> MdbResult<()>

Moves cursor to first entry for key if it exists

fn to_gte_key<'k, K: ToMdbValue<'k>>(&mut self, key: &'k K) -> MdbResult<()>

Moves cursor to first entry for key greater than or equal to ke

fn to_item<'k, 'v, K, V>(&mut self, key: &'k K, value: &'v V) -> MdbResult<()> where K: ToMdbValue<'k>, V: ToMdbValue<'v>

Moves cursor to specific item (for example, if cursor already points to a correct key and you need to delete a specific item through cursor)

fn to_next_key(&mut self) -> MdbResult<()>

Moves cursor to next key, i.e. skip items with duplicate keys

fn to_next_item(&mut self) -> MdbResult<()>

Moves cursor to next item with the same key as current

fn to_prev_key(&mut self) -> MdbResult<()>

Moves cursor to prev entry, i.e. skips items with duplicate keys

fn to_prev_item(&mut self) -> MdbResult<()>

Moves cursor to prev item with the same key as current

fn to_first_item(&mut self) -> MdbResult<()>

Moves cursor to first item with the same key as current

fn to_last_item(&mut self) -> MdbResult<()>

Moves cursor to last item with the same key as current

fn get<T: FromMdbValue<'a>, U: FromMdbValue<'a>>(&mut self) -> MdbResult<(T, U)>

Retrieves current key/value as tuple

fn replace<'a>(&mut self, value: &'a ToMdbValue<'a>) -> MdbResult<()>

Overwrites value for current item Note: overwrites max cur_value.len() bytes

fn add_item<'a>(&mut self, value: &'a ToMdbValue<'a>) -> MdbResult<()>

Adds a new item when created with allowed duplicates

fn del(&mut self) -> MdbResult<()>

Deletes current key

fn del_item(&mut self) -> MdbResult<()>

Deletes only current item

fn del_all(&mut self) -> MdbResult<()>

Deletes all items with same key as current

fn item_count(&self) -> MdbResult<size_t>

Returns count of items with the same key as current