Struct lmdb::core::EnvironmentUnstable

pub struct Environment {
    // some fields omitted
}

Represents LMDB Environment. Should be opened using EnvBuilder

Methods

impl Environment

fn new() -> EnvBuilder

fn stat(&self) -> MdbResult<MDB_stat>

fn info(&self) -> MdbResult<MDB_envinfo>

fn sync(&mut self, force: bool) -> MdbResult<()>

Sync environment to disk

fn set_flags(&mut self, flags: EnvFlags, turn_on: bool) -> MdbResult<()>

This one sets only flags which are available for change even after opening, see also get_flags and get_all_flags

fn get_flags(&self) -> MdbResult<EnvFlags>

Get flags of environment, which could be changed after it was opened use get_all_flags if you need also creation time flags

fn get_all_flags(&self) -> MdbResult<EnvCreateFlags>

Get all flags of environment, including which were specified on creation See also get_flags if you're interested only in modifiable flags

fn get_maxreaders(&self) -> MdbResult<c_uint>

fn get_maxkeysize(&self) -> c_int

fn copy_to_fd(&self, fd: mdb_filehandle_t) -> MdbResult<()>

Creates a backup copy in specified file descriptor

fn get_fd(&self) -> MdbResult<mdb_filehandle_t>

Gets file descriptor of this environment

fn copy_to_path(&self, path: &Path) -> MdbResult<()>

Creates a backup copy in specified path

fn new_transaction(&self) -> MdbResult<Transaction>

Creates a new read-write transaction

Use get_reader to get much faster lock-free alternative

fn get_reader(&self) -> MdbResult<ReadonlyTransaction>

Creates a readonly transaction

fn get_or_create_db<'a>(&'a self, name: &'a str, flags: DbFlags) -> MdbResult<Database>

Returns or creates a named database

Note: set_maxdbis should be called before

fn get_default_db<'a>(&'a self, flags: DbFlags) -> MdbResult<Database>

Returns default database