Struct lmdb::core::NativeTransactionExperimental

pub struct NativeTransaction<'a> {
    // some fields omitted
}

Methods

impl<'a> NativeTransaction<'a>

fn get<'a, T: FromMdbValue<'a>>(&'a self, db: &Database, key: &'a ToMdbValue<'a>) -> MdbResult<T>

fn set<'a>(&'a self, db: &Database, key: &'a ToMdbValue<'a>, value: &'a ToMdbValue<'a>) -> MdbResult<()>

Sets a new value for key, in case of enabled duplicates it actually appends a new value

fn del_item<'a>(&'a self, db: &Database, key: &'a ToMdbValue<'a>, data: &'a ToMdbValue<'a>) -> MdbResult<()>

If duplicate keys are allowed deletes value for key which is equal to data

fn del<'a>(&'a self, db: &Database, key: &'a ToMdbValue<'a>) -> MdbResult<()>

Deletes all values for key

fn new_cursor(&'a self, db: &'a Database) -> MdbResult<Cursor<'a>>

creates a new cursor in current transaction tied to db

fn new_item_iter(&'a self, db: &'a Database, key: &'a ToMdbValue<'a>) -> MdbResult<CursorIterator<'a, CursorItemIter>>

Creates a new item cursor, i.e. cursor which navigates all values with the same key (if AllowsDup was specified)

fn del_db(&self, db: Database) -> MdbResult<()>

Deletes provided database completely

fn empty_db(&self, db: &Database) -> MdbResult<()>

Empties provided database