Struct lmdb::RoTransaction

pub struct RoTransaction<'env> { /* private fields */ }

An LMDB read-only transaction.

Implementations

impl<'env> RoTransaction<'env>

pub fn reset(self) -> InactiveTransaction<'env>

Resets the read-only transaction.

Abort the transaction like Transaction::abort, but keep the transaction handle. InactiveTransaction::renew may reuse the handle. This saves allocation overhead if the process will start a new read-only transaction soon, and also locking overhead if EnvironmentFlags::NO_TLS is in use. The reader table lock is released, but the table slot stays tied to its thread or transaction. Reader locks generally don’t interfere with writers, but they keep old versions of database pages allocated. Thus they prevent the old pages from being reused when writers commit new data, and so under heavy load the database size may grow much more rapidly than otherwise.