Struct lmdb::RwCursor

pub struct RwCursor<'txn> { /* fields omitted */ }

A read-only cursor for navigating items within a database.

Methods

impl<'txn> RwCursor<'txn>

fn put<K, D>(&mut self, key: &K, data: &D, flags: WriteFlags) -> Result<()> where
K: AsRef<[u8]>,
D: AsRef<[u8]>,

Puts a key/data pair into the database. The cursor will be positioned at the new data item, or on failure usually near it.

fn del(&mut self, flags: WriteFlags) -> Result<()>

Deletes the current key/data pair.

Flags

WriteFlags::NO_DUP_DATA may be used to delete all data items for the current key, if the database was opened with DatabaseFlags::DUP_SORT.